Add string publishing and improved error handling in RabbitMQ service
This update includes several changes to better handle different types of data and error scenarios throughout the application. With the addition of the 'PublishString' method in the RabbitMQ service, a broader range of data can be posted. The instantiation of the RabbitMQ service has been made more flexible by allowing the queue name and durability to be set by parameters. These updates also entail consequential changes in the calling methods in the sender and consumer files. Moreover, a completely new 'task' function in a new 'task' file has been added to execute a series of actions upon command line inputs.